+++ tests/probe_endtoend.py
@@ -44,6 +44,10 @@ from http.server import (
 from threading import (
     Thread,
 )
+from socket import (
+    gethostbyname,
+    gethostbyaddr,
+)
 # The probe's output is mixed together with Telepresence output and maybe more
 # output from things like socat or torsocks.  This makes it difficult to
 # extract information from the probe via stdout.  Unfortunately, options apart
@@ -184,6 +188,8 @@ COMMANDS = {
     "probe-url": lambda *urls: list(probe_urls(urls)),
     "probe-also-proxy": probe_also_proxy,
     "disconnect-telepresence": disconnect_telepresence,
+    "gethostbyname": gethostbyname,
+    "gethostbyaddr": gethostbyaddr,
 }